requirements.txt
django==5.0.6
selenium==4.22.0
psycopg2==2.9.9
celery~=5.3.6
djangorestframework==3.15.2
django-celery-beat==2.6.0
django-configurations==2.5
python-dotenv==1.0.1
redis==5.0.7

